[skip] Changed `destroyed?` to `deleted?` to allow saving destroyed_at value (3h)
This was tricky. It seems in recent versions of Rails, the method underlying `save`, `update`, and the like checks if the record is `destroyed?` and returns false if `destroyed?` is true. This means that we would _never_ be able to save a new value for `destroyed_at`, since our custom implementation would always return true, resulting in a failed save. While this change is breaking, it seems necessary to do so, since Houston instances can change their consumption of the model, whereas Rails won't. 😅